+++ src/pgn.rs
@@ -158,24 +158,23 @@ impl PgnLibrary {
     /// }
     /// ```
     pub fn add_entry(&mut self, entry: Entry) -> Result<(), String> {
-        let _id: u32 = match entry {
-            Entry::MessageDefinition(MessageDefinition { ref id, .. }) => id.parse(),
-            Entry::MessageDescription(MessageDescription { ref id, .. }) => id.parse(),
-            Entry::MessageAttribute(MessageAttribute { ref id, .. }) => id.parse(),
+        let _id: u32 = *match entry {
+            Entry::MessageDefinition(MessageDefinition { ref id, .. }) => id,
+            Entry::MessageDescription(MessageDescription { ref id, .. }) => id,
+            Entry::MessageAttribute(MessageAttribute { ref id, .. }) => id,
             Entry::SignalDefinition(..) => {
                 // no id, and by definition must follow MessageDefinition
                 if self.last_id == 0 {
                     return Err("Tried to add SignalDefinition without last ID.".to_string());
                 }
-                Ok(self.last_id)
+                &self.last_id
             }
-            Entry::SignalDescription(SignalDescription { ref id, .. }) => id.parse(),
-            Entry::SignalAttribute(SignalAttribute { ref id, .. }) => id.parse(),
+            Entry::SignalDescription(SignalDescription { ref id, .. }) => id,
+            Entry::SignalAttribute(SignalAttribute { ref id, .. }) => id,
             _ => {
                 return Err(format!("Unsupported entry: {}.", entry).to_string());
             }
-        }
-        .unwrap();
+        };
 
         // CanId{ DP, PF, PS, SA } => Pgn{ PF, PS }
         let pgn = (_id >> 8) & 0x1FFFF;
@@ -348,7 +347,7 @@ impl FromDbc for PgnDefinition {
                 message_len,
                 sending_node,
             }) => {
-                let pgn_long = id.parse::<u32>().unwrap();
+                let pgn_long = id;
                 let pgn = pgn_long & 0x1FFFF;
                 Ok(PgnDefinition::new(
                     pgn,
@@ -364,7 +363,7 @@ impl FromDbc for PgnDefinition {
                 signal_name,
                 description,
             }) => {
-                let pgn_long = id.parse::<u32>().unwrap();
+                let pgn_long = id;
                 let pgn = pgn_long & 0x1FFFF;
                 Ok(PgnDefinition::new(
                     pgn,
@@ -381,7 +380,7 @@ impl FromDbc for PgnDefinition {
                 id,
                 value,
             }) => {
-                let pgn_long = id.parse::<u32>().unwrap();
+                let pgn_long = id;
                 let pgn = pgn_long & 0x1FFFF;
                 Ok(PgnDefinition::new(
                     pgn,
@@ -404,7 +403,7 @@ impl FromDbc for PgnDefinition {
                 message_len,
                 sending_node,
             }) => {
-                let pgn_long = id.parse::<u32>().unwrap();
+                let pgn_long = id;
                 let pgn = pgn_long & 0x1FFFF;
                 self.pgn = pgn;
                 self.pgn_long = pgn_long;
@@ -416,7 +415,7 @@ impl FromDbc for PgnDefinition {
                 signal_name,
                 description,
             }) => {
-                let pgn_long = id.parse::<u32>().unwrap();
+                let pgn_long = id;
                 let pgn = pgn_long & 0x1FFFF;
                 self.pgn = pgn;
                 self.pgn_long = pgn_long;
@@ -429,7 +428,7 @@ impl FromDbc for PgnDefinition {
                 id,
                 value,
             }) => {
-                let pgn_long = id.parse::<u32>().unwrap();
+                let pgn_long = id;
                 let pgn = pgn_long & 0x1FFFF;
                 self.pgn = pgn;
                 self.pgn_long = pgn_long;
@@ -488,7 +487,7 @@ impl FromDbc for PgnDefinition {
 pub struct SpnDefinition {
     name: String,
     pub number: usize,
-    id: String,
+    id: u32,
     description: String,
     start_bit: usize,
     bit_len: usize,
@@ -571,7 +570,7 @@ impl SpnDefinition {
     pub fn new(
         name: String,
         number: usize,
-        id: String,
+        id: u32,
         description: String,
         start_bit: usize,
         bit_len: usize,
@@ -783,7 +782,7 @@ impl From<SignalDefinition> for SpnDefinition {
         SpnDefinition::new(
             name,
             0,
-            "".to_string(),
+            0, // TODO: Some()?
             "".to_string(),
             start_bit,
             bit_len,
@@ -862,7 +861,7 @@ mod tests {
         static ref SPNDEF: SpnDefinition = SpnDefinition::new(
             "Engine_Speed".to_string(),
             190,
-            "2364539904".to_string(),
+            2364539904,
             "A description for Engine speed.".to_string(),
             24,
             16,